samd21,samd51,nrf52840: unify bootloader entry process
