+++ src/machine/machine_atsamd51.go
@@ -1974,9 +1974,9 @@ func (tcc *TCC) Set(channel uint8, value uint32) {
 	}
 }
 
-// ResetProcessor should perform a system reset in preparation
+// EnterBootloader should perform a system reset in preparation
 // to switch to the bootloader to flash new firmware.
-func ResetProcessor() {
+func EnterBootloader() {
 	arm.DisableInterrupts()
 
 	// Perform magic reset into bootloader, as mentioned in
