+++ src/machine/machine_nrf52840_usb_reset_none.go
@@ -3,5 +3,7 @@
 
 package machine
 
-func checkShouldReset() {
+// EnterBootloader resets the chip into the serial bootloader.
+func EnterBootloader() {
+	// skip
 }
